refactor: update product list routes
