+++ src/pages/products/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductList } from './product-list';
+export { default as Upload } from './upload';
