+++ src/pages/products/pages/upload/index.tsx
@@ -1,0 +1,11 @@
+import { AppBarBack } from '@/assets/icons';
+import { AppBar, Layout } from '@/components';
+
+export default function Upload() {
+  return (
+    <Layout>
+      <AppBar prefix={<AppBarBack height="24px" />} />
+      Upload
+    </Layout>
+  );
+}
